+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix of #701 - editors/number.js and editors/integer.js don't change values when validation is failed
 - Fix of #716 - add ignore for allOf to fall in line with existing ignores of anyOf/oneOf for additionalProperties validation
 - Fix of #714 - Checkboxes inside object tables duplicate labels from heading
+- Added copy button to arrays in table format 
 
 ### 2.0.0-dev
   - Fix of #643 - Allow use of themes not compiled directly into the build